bugfix Выбор места работы

+++ app/src/main/java/ru/practicum/android/diploma/filters/choosearea/ui/ChooseAreaFragment.kt
@@ -4,6 +4,8 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import android.widget.ImageView
+import androidx.core.content.ContextCompat
 import androidx.core.view.isVisible
 import androidx.fragment.app.Fragment
 import androidx.navigation.fragment.findNavController
@@ -32,17 +34,15 @@ class ChooseAreaFragment : Fragment() {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
-        viewModelChooseArea.areaSettings.observe(viewLifecycleOwner) { area ->
-            renderAreaSettings()
-        }
+        renderAreaSettings()
 
-        binding.countryEd.setOnClickListener {
+        binding.countryCl.setOnClickListener {
             findNavController().navigate(
                 R.id.action_chooseAreaFragment_to_chooseCountryFragment
             )
         }
 
-        binding.regionEd.setOnClickListener {
+        binding.regionCl.setOnClickListener {
             findNavController().navigate(
                 R.id.action_chooseAreaFragment_to_chooseRegionFragment,
                 ChooseRegionFragment.createArgs(areaSettings?.countryInfo?.id)
@@ -50,21 +50,27 @@ class ChooseAreaFragment : Fragment() {
         }
 
         binding.countryArrowAndCleanIv.setOnClickListener {
-            viewModelChooseArea.deleteCountrySettings()
-            with(binding) {
+            if (checkIvImage(binding.countryArrowAndCleanIv, R.drawable.clear_24px_input_edittext_button)) {
+                viewModelChooseArea.deleteCountrySettings()
                 areaSettings = null
-                countryArrowAndCleanIv.setImageResource(R.drawable.arrow_forward_24px_button)
-                regionArrowAndCleanIv.setImageResource(R.drawable.arrow_forward_24px_button)
-                countryEd.text = null
-                regionEd.text = null
-                applyBt.isVisible = false
+                hideRegion()
+                hideCountry()
+            } else {
+                findNavController().navigate(
+                    R.id.action_chooseAreaFragment_to_chooseCountryFragment
+                )
             }
         }
+
         binding.regionArrowAndCleanIv.setOnClickListener {
-            viewModelChooseArea.saveAreaSettings(AreaInfo("", "", areaSettings!!.countryInfo))
-            with(binding) {
-                regionArrowAndCleanIv.setImageResource(R.drawable.arrow_forward_24px_button)
-                regionEd.text = null
+            if (checkIvImage(binding.regionArrowAndCleanIv, R.drawable.clear_24px_input_edittext_button)) {
+                viewModelChooseArea.saveAreaSettings(AreaInfo("", "", areaSettings!!.countryInfo))
+                hideRegion()
+            } else {
+                findNavController().navigate(
+                    R.id.action_chooseAreaFragment_to_chooseRegionFragment,
+                    ChooseRegionFragment.createArgs(areaSettings?.countryInfo?.id)
+                )
             }
         }
 
@@ -73,7 +79,6 @@ class ChooseAreaFragment : Fragment() {
         }
 
         binding.arrowBackIv.setOnClickListener {
-            viewModelChooseArea.deleteCountrySettings()
             findNavController().navigateUp()
         }
     }
@@ -86,20 +91,60 @@ class ChooseAreaFragment : Fragment() {
     private fun renderAreaSettings() {
         areaSettings = viewModelChooseArea.getAreaSettings()
         if (areaSettings != null) {
-            with(binding) {
-                countryEd.setText(areaSettings!!.countryInfo.name)
-                applyBt.isVisible = true
-                countryArrowAndCleanIv.isVisible = true
-                countryArrowAndCleanIv.setImageResource(R.drawable.clear_24px_input_edittext_button)
-            }
+            showCountry()
             if (areaSettings!!.name.isNotEmpty()) {
-                with(binding) {
-                    regionEd.setText(areaSettings!!.name)
-                    regionArrowAndCleanIv.isVisible = true
-                    regionArrowAndCleanIv.setImageResource(R.drawable.clear_24px_input_edittext_button)
-                }
-
+                showRegion()
+            } else {
+                hideRegion()
             }
+        } else {
+            hideCountry()
         }
+    }
+
+    private fun showCountry() {
+        with(binding) {
+            countryTil.isVisible = true
+            countryTv.isVisible = false
+            countryDarkTv.isVisible = true
+            countryDarkTv.text = areaSettings!!.countryInfo.name
+            applyBt.isVisible = true
+            countryArrowAndCleanIv.setImageResource(R.drawable.clear_24px_input_edittext_button)
+        }
+    }
+
+    private fun hideCountry() {
+        with(binding) {
+            countryTv.isVisible = true
+            countryDarkTv.isVisible = false
+            countryTil.isVisible = false
+            applyBt.isVisible = false
+            countryArrowAndCleanIv.setImageResource(R.drawable.arrow_forward_24px_button)
+        }
+    }
+
+    private fun showRegion() {
+        with(binding) {
+            regionTil.isVisible = true
+            regionTv.isVisible = false
+            regionDarkTv.isVisible = true
+            regionDarkTv.text = areaSettings!!.name
+            regionArrowAndCleanIv.setImageResource(R.drawable.clear_24px_input_edittext_button)
+        }
+    }
+
+    private fun hideRegion() {
+        with(binding) {
+            regionTv.isVisible = true
+            regionDarkTv.isVisible = false
+            regionTil.isVisible = false
+            regionArrowAndCleanIv.setImageResource(R.drawable.arrow_forward_24px_button)
+        }
+    }
+
+    private fun checkIvImage(imageView: ImageView, imageResourceId: Int): Boolean {
+        return imageView.drawable?.let {
+            return (it.constantState == ContextCompat.getDrawable(imageView.context, imageResourceId)?.constantState)
+        } ?: false
     }
 }
